requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pyspark==3.5.0
py4j==0.10.9.7
pandas==2.1.4
dbt-core==1.7.0
dbt-postgres==1.7.0
apache-airflow==2.8.0
pytest==7.4.3
pytest-mock==3.12.0
